requirements.txt
flask
flask-cors
requests
psycopg2-binary